+++ src/main/java/com/example/travelDiary/application/events/travel/ScheduleCreatedEvent.java
@@ -1,6 +1,6 @@
 package com.example.travelDiary.application.events.travel;
 
-import com.example.travelDiary.domain.model.travel.schedule.Schedule;
+import com.example.travelDiary.domain.model.travel.Schedule;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
