+++ src/main/java/com/example/travelDiary/application/service/travel/TravelPlanAccessService.java
@@ -1,7 +1,6 @@
 package com.example.travelDiary.application.service.travel;
 
-import com.example.travelDiary.application.service.travel.schedule.ScheduleEventMediator;
-import com.example.travelDiary.application.service.travel.schedule.ScheduleMutationService;
+import com.example.travelDiary.application.service.travel.schedule.ScheduleQueryService;
 import com.example.travelDiary.domain.model.travel.TravelPlan;
 import com.example.travelDiary.domain.model.wallet.aggregate.MonetaryEvent;
 import com.example.travelDiary.repository.persistence.travel.TravelPlanRepository;
@@ -20,15 +19,13 @@ import java.util.UUID;
 public class TravelPlanAccessService {
     private final TravelPlanRepository travelPlanRepository;
     private final ConversionService conversionService;
-    private final ScheduleMutationService scheduleMutationService;
-    private final ScheduleEventMediator scheduleEventMediator;
+    private final ScheduleQueryService scheduleQueryService;
 
     @Autowired
-    public TravelPlanAccessService(TravelPlanRepository travelPlanRepository, ConversionService conversionService, ScheduleMutationService scheduleMutationService, ScheduleEventMediator scheduleEventMediator) {
+    public TravelPlanAccessService(TravelPlanRepository travelPlanRepository, ConversionService conversionService, ScheduleQueryService scheduleQueryService) {
         this.travelPlanRepository = travelPlanRepository;
         this.conversionService = conversionService;
-        this.scheduleMutationService = scheduleMutationService;
-        this.scheduleEventMediator = scheduleEventMediator;
+        this.scheduleQueryService = scheduleQueryService;
     }
 
     public Page<TravelPlan> getTravelPageContainingName(String name, int pageNumber, int pageSize) {
@@ -68,7 +65,7 @@ public class TravelPlanAccessService {
                 .orElseThrow()
                 .getScheduleList()
                 .stream()
-                .flatMap(schedule -> scheduleEventMediator
+                .flatMap(schedule -> scheduleQueryService
                         .getAssociatedMonetaryEvent(
                                 schedule
                                         .getId()
