+++ src/main/java/com/example/travelDiary/application/service/travel/schedule/ScheduleMutationService.java
@@ -4,11 +4,11 @@ import com.example.travelDiary.application.events.EventPublisher;
 import com.example.travelDiary.application.events.travel.ScheduleCreatedEvent;
 import com.example.travelDiary.application.events.travel.ScheduleDeletedEvent;
 import com.example.travelDiary.application.events.travel.ScheduleUpdatedEvent;
-import com.example.travelDiary.application.service.travel.place.PlaceMutationService;
+import com.example.travelDiary.application.service.location.PlaceMutationService;
 import com.example.travelDiary.application.service.travel.RouteAccessService;
 import com.example.travelDiary.domain.model.location.Place;
 import com.example.travelDiary.domain.model.travel.Route;
-import com.example.travelDiary.domain.model.travel.schedule.Schedule;
+import com.example.travelDiary.domain.model.travel.Schedule;
 import com.example.travelDiary.domain.model.travel.TravelPlan;
 import com.example.travelDiary.repository.persistence.travel.ScheduleRepository;
 import com.example.travelDiary.repository.persistence.travel.TravelPlanRepository;
