+++ src/main/java/com/example/travelDiary/domain/model/travel/TravelPlan.java
@@ -1,6 +1,5 @@
 package com.example.travelDiary.domain.model.travel;
 
-import com.example.travelDiary.domain.model.travel.schedule.Schedule;
 import jakarta.persistence.*;
 import lombok.Data;
 import org.hibernate.annotations.Cascade;
